Move "loading" properties to a new "source" layer type
