+++ debug/loading-pattern.html
@@ -14,7 +14,7 @@
 <body>
 <div id='map'></div>
 
-<script src='mapbox-gl-dev.js'></script>
+<script src='/dist/mapbox-gl-dev.js'></script>
 <script>
 
 mapboxgl.accessToken = 'pk.eyJ1IjoibHVjYXN3b2oiLCJhIjoiNWtUX3JhdyJ9.WtCTtw6n20XV2DwwJHkGqQ';
@@ -35,13 +35,15 @@ map.on('load', function() {
     });
 
     map.addLayer({
-        "id": "raster-layer",
-        "type": "raster",
-        "source": "raster-source",
+        "id": "raster-source-layer",
+        "type": "source",
+        "source": "composite",
         "paint": {
-            "raster-loading-pattern": "rocket-15"
+            'source-loading-pattern': 'circle-11',
+            'source-loading-opacity': 0.5
         }
     });
+
 });
 
 </script>
