+++ package.json
@@ -23,7 +23,7 @@
     "grid-index": "^1.0.0",
     "mapbox-gl-function": "mapbox/mapbox-gl-function#41c6724e2bbd7bd1eb5991451bbf118b7d02b525",
     "mapbox-gl-shaders": "mapbox/mapbox-gl-shaders#94518afb141b376e7f86e492c1b022c0159ca4a3",
-    "mapbox-gl-style-spec": "mapbox/mapbox-gl-style-spec#1bfcc827f900b074c5ba846ebb8d2ca9c9a572a7",
+    "mapbox-gl-style-spec": "mapbox/mapbox-gl-style-spec#781999ca90bb52b0d56a6e9dfe123e4ca352c753",
     "mapbox-gl-supported": "^1.2.0",
     "package-json-versionify": "^1.0.2",
     "pbf": "^1.3.2",
